requirements.txt
torch
torchvision
torchdiffeq
matplotlib
scipy
numpy
